requirements.txt
absl-py==2.1.0
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
astunparse==1.6.3
attrs==23.2.0
audio-recorder-streamlit==0.0.8
audioread==3.0.1
besser-bot-framework==1.2.1
blinker==1.8.2
blis==0.7.11
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.5
cymem==2.0.8
dateparser==1.1.8
decorator==5.1.1
distro==1.9.0
filelock==3.15.1
flatbuffers==24.3.25
fsspec==2024.6.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.30.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.25.2
huggingface-hub==0.23.4
idna==3.7
importlib-metadata==6.11.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
keras==2.14.0
langcodes==3.4.0
language_data==1.2.0
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.1
llvmlite==0.43.0
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.2.0
msgpack==1.0.8
murmurhash==1.0.10
nltk==3.8.1
numba==0.60.0
numpy==1.26.1
oauthlib==3.2.2
openai==1.8.0
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.1
pillow==10.3.0
platformdirs==4.2.2
plotly==5.18.0
pooch==1.8.2
preshed==3.0.9
protobuf==4.25.3
psycopg2-binary==2.9.9
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==20.6
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
replicate==0.23.1
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
snowballstemmer==2.2.0
soundfile==0.12.1
soxr==0.3.7
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.10.0
SQLAlchemy==2.0.29
srsly==2.4.8
streamlit==1.27.2
streamlit-antd-components==0.3.2
tenacity==8.3.0
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.37.0
tensorflow-macos==2.14.0
termcolor==2.4.0
text2num==2.5.0
thinc==8.2.4
threadpoolctl==3.5.0
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
transformers==4.37.0
typer==0.9.4
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
validators==0.28.3
wasabi==1.1.3
weasel==0.3.4
websocket-client==1.6.4
websockets==11.0.3
Werkzeug==3.0.3
wrapt==1.14.1
zipp==3.19.2
